[frontend] Reorganize folder structure
