+++ frontend/src/App.vue
@@ -12,11 +12,11 @@
 <script lang="ts">
 import Vue from 'vue'
 import Component from 'vue-class-component'
-import NavigationBar from './components/NavigationBar.vue'
-import Snackbar from './components/Snackbar.vue'
+import NavigationBar from './components/misc/NavigationBar.vue'
+import Snackbar from './components/misc/Snackbar.vue'
 import { vxm } from './store'
 import { Watch } from 'vue-property-decorator'
-import ThemeSelector from './components/ThemeSelector.vue'
+import ThemeSelector from './components/misc/ThemeSelector.vue'
 import { storeToLocalStorage } from './store/persistence'
 import { Route } from 'vue-router'
 import '@/css/AnsiTheme.css'
