+++ frontend/src/components/graphs/detail/DimensionSelection.vue
