+++ frontend/src/components/graphs/detail/MatrixDimensionSelection.vue
