+++ frontend/src/components/graphs/graph/EchartsDetailGraph.vue
