+++ frontend/src/components/graphs/helper/GraphSettings.vue
