+++ frontend/src/components/graphs/helper/GraphTimespanControls.vue
