+++ frontend/src/components/graphs/helper/ShareGraphLinkDialog.vue
