+++ frontend/src/components/misc/InlineMinimalRepoDisplay.vue
