+++ frontend/src/components/misc/NavigationBar.vue
@@ -98,8 +98,8 @@ import Vue from 'vue'
 import Component from 'vue-class-component'
 import { VuetifyIcon } from 'vuetify/types/services/icons'
 import VueRouterEx from 'vue-router/types/router'
-import router from '../router'
-import LoginDialog from '../components/dialogs/LoginDialog.vue'
+import router from '../../router'
+import LoginDialog from './LoginDialog.vue'
 import { mdiAccountCircleOutline, mdiLogout } from '@mdi/js'
 import { vxm } from '@/store'
 import { Watch } from 'vue-property-decorator'
