+++ frontend/src/components/overviews/TarTaskOverview.vue
@@ -54,8 +54,8 @@ import Vue from 'vue'
 import Component from 'vue-class-component'
 import { Prop } from 'vue-property-decorator'
 import { TarTaskSource, RunId, TaskId } from '@/store/types'
-import CommitChip from '../CommitChip.vue'
-import InlineMinimalRepoNameDisplay from '../InlineMinimalRepoDisplay.vue'
+import CommitChip from '../runs/CommitChip.vue'
+import InlineMinimalRepoNameDisplay from '../misc/InlineMinimalRepoDisplay.vue'
 import { RawLocation } from 'vue-router'
 import { mdiScaleBalance } from '@mdi/js'
