+++ frontend/src/components/queue/QueueOverview.vue
@@ -87,8 +87,8 @@ import Component from 'vue-class-component'
 import { vxm } from '@/store'
 import { Task, Worker } from '@/store/types'
 import { mdiDelete, mdiRocket } from '@mdi/js'
-import CommitOverviewBase from './CommitOverviewBase.vue'
-import TarTaskOverview from './TarTaskOverview.vue'
+import CommitOverviewBase from '../overviews/CommitOverviewBase.vue'
+import TarTaskOverview from '../overviews/TarTaskOverview.vue'
 import { formatDurationHuman } from '@/util/TimeUtil'
 import { RawLocation } from 'vue-router'
