+++ frontend/src/components/queue/WorkerOverview.vue
