+++ frontend/src/components/rundetail/CommitDetail.vue
@@ -123,8 +123,8 @@ import Component from 'vue-class-component'
 import { Prop } from 'vue-property-decorator'
 import { Commit, RunId, TrackedCommitDescription } from '@/store/types'
 import { formatDateUTC, formatDate } from '@/util/TimeUtil'
-import InlineMinimalRepoNameDisplay from '../InlineMinimalRepoDisplay.vue'
-import CommitBenchmarkActions from '../CommitBenchmarkActions.vue'
+import InlineMinimalRepoNameDisplay from '../misc/InlineMinimalRepoDisplay.vue'
+import CommitBenchmarkActions from '../runs/CommitBenchmarkActions.vue'
 import CommitNavigationButton from './CommitNavigationButton.vue'
 import { mdiCompassOffOutline } from '@mdi/js'
