+++ frontend/src/components/search/SearchResultTar.vue
@@ -40,8 +40,8 @@ import Vue from 'vue'
 import Component from 'vue-class-component'
 import { Prop } from 'vue-property-decorator'
 import { RawLocation } from 'vue-router'
-import CommitChip from '@/components/CommitChip.vue'
-import InlineMinimalRepoDisplay from '@/components/InlineMinimalRepoDisplay.vue'
+import CommitChip from '@/components/runs/CommitChip.vue'
+import InlineMinimalRepoDisplay from '@/components/misc/InlineMinimalRepoDisplay.vue'
 import { mdiFolderZipOutline } from '@mdi/js'
 import { SearchItemRun } from '@/store/types'
