+++ frontend/src/views/RepoDetailFrame.vue
@@ -49,9 +49,9 @@ import Vue from 'vue'
 import { Component } from 'vue-property-decorator'
 import { vxm } from '@/store'
 import { Repo } from '@/store/types'
-import RepoSelectionComponent from '../components/RepoSelectionComponent.vue'
+import RepoSelectionComponent from '../components/misc/RepoSelectionComponent.vue'
 import { mdiPlusCircleOutline } from '@mdi/js'
-import RepoAddDialog from '../components/dialogs/RepoAddDialog.vue'
+import RepoAddDialog from '../components/repodetail/RepoAddDialog.vue'
 import NotFound404 from './NotFound404.vue'
 import RepoBaseInformation from '@/components/repodetail/RepoBaseInformation.vue'
 import RepoGraphView from '@/views/RepoGraphView.vue'
